relay/pyproject.toml
[build-system]
requires      = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name    = "peerdup-relay"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "pynacl>=1.5",
    "base58>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[project.scripts]
peerdup-relay = "relay.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["relay*"]
